updating watch
