+++ packages/signals/lib/signals_flutter.dart
@@ -5,6 +5,5 @@ export 'src/extensions/signal.dart';
 export 'src/extensions/value_listenable.dart';
 export 'src/extensions/value_notifier.dart';
 export 'src/widgets/watch.dart';
-export 'src/widgets/watch_builder.dart';
 export 'src/widgets/signal_widget.dart';
 export 'src/watch_signal.dart';
