+++ website/src/content/docs/examples/flutter-counter.md
@@ -72,7 +72,7 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: Text(widget.title),
         actions: [
-          Watch(() {
+          Watch((_) {
             final isDark = brightness() == Brightness.dark;
             return IconButton(
               onPressed: () {
@@ -90,7 +90,7 @@ class _MyHomePageState extends State<MyHomePage> {
             const Text(
               'You have pushed the button this many times:',
             ),
-            Watch(() {
+            Watch((context) {
               return Text(
                 '$counter',
                 style: Theme.of(context).textTheme.headlineMedium!,
